renamed table_queries to datatables for a better understanding
